requirements.txt
Django==3.2.15
djangorestframework==3.14.0
django-debug-toolbar==3.2.4
django-cache-url==3.4.5
dj-database-url==2.2.0
dj-email-url==1.0.6
django-phonenumber-field==7.1.0
phonenumbers==9.0.5
Pillow==8.2.0
requests==2.32.3
geopy==2.4.1
rollbar==1.3.0
environs==9.3.2
marshmallow==3.19.0
gunicorn>=20.1.0
